# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateOrderRequest:
    properties:
      products:
        items:
          $ref: '#/definitions/dto.OrderProduct'
        type: array
    required:
    - products
    type: object
  dto.CreateOrderResponse:
    properties:
      id:
        type: string
      status:
        type: string
      total_amount:
        type: number
      user_id:
        type: string
    type: object
  dto.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    required:
    - description
    - name
    - price
    - stock
    type: object
  dto.CreateProductResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
  dto.GetOrderResponse:
    properties:
      id:
        type: string
      products:
        items:
          $ref: '#/definitions/models.OrderProduct'
        type: array
      status:
        type: string
      total_amount:
        type: number
      user_id:
        type: string
    type: object
  dto.GetProductResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
  dto.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginUserResponse:
    properties:
      access_token:
        type: string
      aceess_token_expires_at:
        type: integer
      refresh_token:
        type: string
    type: object
  dto.OrderProduct:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterUserResponse:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
  dto.UpdateOrderStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  dto.UpdateOrderStatusResponse:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
  dto.UpdateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
  dto.UpdateProductResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
  helpers.ApiResponse:
    properties:
      data:
        description: The data returned by the API
      message:
        description: A message providing additional information about the response
        type: string
      success:
        description: Indicates whether the request was successful
        type: boolean
    type: object
  models.OrderProduct:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
info:
  contact:
    email: support@petmanagement.com
    name: API Support Team
  description: A RESTful API for an e-commerce application. This API will handle basic
    CRUD operations for products and orders, and provide user management and authentication.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Shopper Management API
  version: "1.0"
paths:
  /orders:
    post:
      consumes:
      - application/json
      description: Create a new order
      parameters:
      - description: Order create request
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateOrderResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Create a new order
      tags:
      - orders
  /orders/{id}/cancel:
    put:
      description: Cancel an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Cancel an order
      tags:
      - orders
  /orders/{id}/status:
    put:
      description: Update the status of an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order status update request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateOrderStatusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Update order status
      tags:
      - admin
  /orders/me:
    get:
      description: Get all orders for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.GetOrderResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Get all orders for authenticated user
      tags:
      - orders
  /products:
    get:
      description: Get all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.GetProductResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Get all products
      tags:
      - products
      - admin
    post:
      consumes:
      - application/json
      description: Create a new product with name, description, and price
      parameters:
      - description: Product creation request
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Create a new product
      tags:
      - admin
  /products/{id}:
    delete:
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Delete a product
      tags:
      - admin
    get:
      description: Get a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetProductResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Get a product
      tags:
      - products
      - admin
    put:
      consumes:
      - application/json
      description: Update a product's name, description, and price
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product update request
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Update a product
      tags:
      - admin
  /users:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User registration request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterUserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      summary: Register a new user
      tags:
      - users
  /users/auth:
    post:
      consumes:
      - application/json
      description: Login a user with email and password
      parameters:
      - description: User login request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginUserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      summary: Login a user
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  JWT:
    description: JWT security accessToken. Please add it in the format "Bearer {AccessToken}"
      to authorize your requests.
    in: header
    name: Authorization
    scopes:
      admin: Grants read and write access to administrative information
    type: apiKey
swagger: "2.0"